Month Dialog first stepto replace
